requirements.txt
# Core AI/ML libraries
torch==2.5.1+cu121
transformers==4.57.1
accelerate==1.11.0
peft==0.17.1
bitsandbytes==0.48.2

# ModelScope (for Qwen models)
modelscope==1.32.0

# Web framework for API
fastapi==0.121.3
uvicorn==0.38.0

# Data processing
datasets==3.6.0

# Swift framework (for advanced training)
ms-swift==3.10.2

# HTTP client for testing
requests==2.32.5

# Other utilities
